+++ services/ecs/struct_command.go
@@ -23,7 +23,7 @@ type Command struct {
 	Description     string         `json:"Description" xml:"Description"`
 	CommandContent  string         `json:"CommandContent" xml:"CommandContent"`
 	WorkingDir      string         `json:"WorkingDir" xml:"WorkingDir"`
-	Timeout         int            `json:"Timeout" xml:"Timeout"`
+	Timeout         int64          `json:"Timeout" xml:"Timeout"`
 	CreationTime    string         `json:"CreationTime" xml:"CreationTime"`
 	EnableParameter bool           `json:"EnableParameter" xml:"EnableParameter"`
 	ParameterNames  ParameterNames `json:"ParameterNames" xml:"ParameterNames"`
