+++ services/ecs/struct_instance_type.go
@@ -19,7 +19,7 @@ package ecs
 type InstanceType struct {
 	MemorySize                  float64 `json:"MemorySize" xml:"MemorySize"`
 	EniPrivateIpAddressQuantity int     `json:"EniPrivateIpAddressQuantity" xml:"EniPrivateIpAddressQuantity"`
-	InstancePpsRx               int     `json:"InstancePpsRx" xml:"InstancePpsRx"`
+	InstancePpsRx               int64   `json:"InstancePpsRx" xml:"InstancePpsRx"`
 	CpuCoreCount                int     `json:"CpuCoreCount" xml:"CpuCoreCount"`
 	Cores                       int     `json:"Cores" xml:"Cores"`
 	Memory                      int     `json:"Memory" xml:"Memory"`
@@ -33,10 +33,10 @@ type InstanceType struct {
 	SupportIoOptimized          string  `json:"SupportIoOptimized" xml:"SupportIoOptimized"`
 	InstanceTypeFamily          string  `json:"InstanceTypeFamily" xml:"InstanceTypeFamily"`
 	InitialCredit               int     `json:"InitialCredit" xml:"InitialCredit"`
-	InstancePpsTx               int     `json:"InstancePpsTx" xml:"InstancePpsTx"`
+	InstancePpsTx               int64   `json:"InstancePpsTx" xml:"InstancePpsTx"`
 	LocalStorageAmount          int     `json:"LocalStorageAmount" xml:"LocalStorageAmount"`
 	InstanceFamilyLevel         string  `json:"InstanceFamilyLevel" xml:"InstanceFamilyLevel"`
-	LocalStorageCapacity        int     `json:"LocalStorageCapacity" xml:"LocalStorageCapacity"`
+	LocalStorageCapacity        int64   `json:"LocalStorageCapacity" xml:"LocalStorageCapacity"`
 	GPUSpec                     string  `json:"GPUSpec" xml:"GPUSpec"`
 	LocalStorageCategory        string  `json:"LocalStorageCategory" xml:"LocalStorageCategory"`
 	InstanceBandwidthTx         int     `json:"InstanceBandwidthTx" xml:"InstanceBandwidthTx"`
